hirehub/requirements.txt
livekit-agents[openai]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv>=1.0.0
httpx>=0.25.0 